# server/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Laravel",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/download/downloadSong": {
            "post": {
                "operationId": "download.downloadSong",
                "summary": "\u5355\u66f2\u4e0b\u8f7d\uff1abody \u4e3a\u641c\u7d22\u8fd4\u56de\u7684\u5b8c\u6574\u6b4c\u66f2\u8bb0\u5f55\uff0cbrType \u7701\u7565\u65f6 worker \u81ea\u52a8\u9009\u6700\u9ad8\u97f3\u8d28",
                "tags": [
                    "Download"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "plugName": {
                                        "type": "string"
                                    },
                                    "brType": {
                                        "type": [
                                            "string",
                                            "null"
                                        ]
                                    }
                                },
                                "required": [
                                    "id",
                                    "name",
                                    "plugName"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "const": 200
                                                },
                                                "msg": {
                                                    "type": "null"
                                                },
                                                "data": {
                                                    "type": "null"
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "msg",
                                                "data"
                                            ]
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "const": 500
                                                },
                                                "msg": {
                                                    "type": "string",
                                                    "const": "\u6b4c\u66f2\u8bb0\u5f55\u7f3a\u5c11 id/name"
                                                },
                                                "data": {
                                                    "type": "null"
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "msg",
                                                "data"
                                            ]
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "const": 500
                                                },
                                                "msg": {
                                                    "type": "string"
                                                },
                                                "data": {
                                                    "type": "null"
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "msg",
                                                "data"
                                            ]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/download/downloadAlbum": {
            "post": {
                "operationId": "download.downloadAlbum",
                "summary": "\u6574\u5f20\u4e13\u8f91\u4e0b\u8f7d\uff1abody \u4e3a\u4e13\u8f91\u8bb0\u5f55 + \u53ef\u9009 bit\uff08\u6574\u6570\u7801\u7387\uff0c\u7701\u7565=\u81ea\u52a8\u6700\u9ad8\uff09\u3002\n\u540c\u6b65\u5c55\u5f00\u66f2\u76ee\uff08\u4e00\u6b21\u4e0a\u6e38\u8bf7\u6c42\uff09\u5e76\u8fd4\u56de\u4efb\u52a1\u6570\u7ec4\uff0c\u524d\u7aef\u53d6\u6570\u7ec4\u957f\u5ea6\u505a\u8ba1\u6570\u63d0\u793a",
                "tags": [
                    "Download"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "albumid": {
                                        "type": "string"
                                    },
                                    "plugName": {
                                        "type": "string"
                                    },
                                    "bit": {
                                        "type": [
                                            "integer",
                                            "null"
                                        ]
                                    }
                                },
                                "required": [
                                    "albumid",
                                    "plugName"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "const": 200
                                                },
                                                "msg": {
                                                    "type": "null"
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "downloadGid": {
                                                                "type": [
                                                                    "string",
                                                                    "null"
                                                                ]
                                                            },
                                                            "downloadTime": {
                                                                "type": [
                                                                    "string",
                                                                    "null"
                                                                ]
                                                            },
                                                            "downloadFile": {
                                                                "type": [
                                                                    "string",
                                                                    "null"
                                                                ]
                                                            },
                                                            "downloadMusicId": {
                                                                "type": "string"
                                                            },
                                                            "downloadPlugName": {
                                                                "type": "string"
                                                            },
                                                            "downloadBrType": {
                                                                "type": "string"
                                                            },
                                                            "downloadMusicname": {
                                                                "type": "string"
                                                            },
                                                            "downloadArtistname": {
                                                                "type": [
                                                                    "string",
                                                                    "null"
                                                                ]
                                                            },
                                                            "downloadAlbumname": {
                                                                "type": [
                                                                    "string",
                                                                    "null"
                                                                ]
                                                            },
                                                            "downloadMsg": {
                                                                "type": [
                                                                    "string",
                                                                    "null"
                                                                ]
                                                            },
                                                            "downloadMusicInfo": {
                                                                "type": [
                                                                    "string",
                                                                    "null"
                                                                ]
                                                            },
                                                            "downloadStatus": {
                                                                "type": "string"
                                                            },
                                                            "springName": {
                                                                "type": "null"
                                                            },
                                                            "audioBook": {
                                                                "type": "null"
                                                            },
                                                            "downloadUpdateTime": {
                                                                "type": [
                                                                    "string",
                                                                    "null"
                                                                ]
                                                            },
                                                            "rewriteMp3tag": {
                                                                "type": "null"
                                                            },
                                                            "downloadBits": {
                                                                "type": "null"
                                                            },
                                                            "downloadBrTypes": {
                                                                "type": [
                                                                    "array",
                                                                    "null"
                                                                ],
                                                                "items": {}
                                                            }
                                                        },
                                                        "required": [
                                                            "id",
                                                            "downloadGid",
                                                            "downloadTime",
                                                            "downloadFile",
                                                            "downloadMusicId",
                                                            "downloadPlugName",
                                                            "downloadBrType",
                                                            "downloadMusicname",
                                                            "downloadArtistname",
                                                            "downloadAlbumname",
                                                            "downloadMsg",
                                                            "downloadMusicInfo",
                                                            "downloadStatus",
                                                            "springName",
                                                            "audioBook",
                                                            "downloadUpdateTime",
                                                            "rewriteMp3tag",
                                                            "downloadBits",
                                                            "downloadBrTypes"
                                                        ]
                                                    }
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "msg",
                                                "data"
                                            ]
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "const": 500
                                                },
                                                "msg": {
                                                    "type": "string"
                                                },
                                                "data": {
                                                    "type": "null"
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "msg",
                                                "data"
                                            ]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/download/downloadArtistAlbum": {
            "post": {
                "operationId": "download.downloadArtistAlbum",
                "summary": "\u6b4c\u624b\u5168\u90e8\u4e13\u8f91\u4e0b\u8f7d\uff1a\u4e13\u8f91\u591a\uff08\u6bcf\u5f20\u4e00\u6b21\u4e0a\u6e38\u8bf7\u6c42\uff09\uff0c\u5165\u961f\u5f02\u6b65\u5c55\u5f00\uff0c\u4efb\u52a1\u5728 task/list \u4e2d\u6e10\u8fdb\u51fa\u73b0",
                "tags": [
                    "Download"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "artistid": {
                                        "type": "string"
                                    },
                                    "plugName": {
                                        "type": "string"
                                    },
                                    "bit": {
                                        "type": [
                                            "integer",
                                            "null"
                                        ]
                                    }
                                },
                                "required": [
                                    "artistid",
                                    "plugName"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "const": 200
                                                },
                                                "msg": {
                                                    "type": "null"
                                                },
                                                "data": {
                                                    "type": "null"
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "msg",
                                                "data"
                                            ]
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "integer",
                                                    "const": 500
                                                },
                                                "msg": {
                                                    "type": "string"
                                                },
                                                "data": {
                                                    "type": "null"
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "msg",
                                                "data"
                                            ]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            }
        }
    }
}